Implement cinder-volume-hyperv ansible role
Implement an ansible role that will deploy cinder-volume as a cinder
volume smb backend on HyperV compute nodes.

This will install and configure the cinder-volume on the HyperV host.
cinder-volume is also capable to handle iSCSI backend, but only
smb backed is configured with this role. This role does not
handle the creation of the SMB share, so this needs to be
already configured.
https://docs.openstack.org/ocata/config-reference/block-storage/drivers/smbfs-volume-driver.html

Change-Id: Ie19a8f578cb1c63739c88a163ea465e420d6fa14
Implements: blueprint cinder-smb-hyperv-ansible-role